pytest: Separate device tests
